added a PythonChallenge

+++ pwnshop/__main__.py
@@ -231,7 +231,7 @@ def handle_apply(args):
                     shutil.copy2(challenge.src_path, os.path.join(out_dir, os.path.basename(challenge.src_path)))
                 print(f"... copying binary {os.path.basename(challenge.bin_path)} to {out_dir}")
                 shutil.copy2(challenge.bin_path, os.path.join(out_dir, os.path.basename(challenge.bin_path)))
-                if os.path.exists(challenge.lib_path):
+                if hasattr(challenge, "lib_path") and os.path.exists(challenge.lib_path):
                     print(f"... copying libraries {os.path.basename(challenge.lib_path)} to {out_dir}")
                     shutil.copytree(challenge.lib_path, os.path.join(out_dir, os.path.basename(challenge.lib_path)), dirs_exist_ok=True)
